+++ api/package.json
@@ -3,8 +3,8 @@
   "version": "1.0.0",
   "type": "module",
   "scripts": {
-    "dev": "nodemon src/server.js",
-    "start": "node src/server.js",
+    "dev": "nodemon src/server.cjs",
+    "start": "node src/server.cjs",
     "test": "echo 'add tests'"
   },
   "devDependencies": {
